25. simple_network_monitor_dashboard/requirements.txt
Flask
pysnmp
pyyaml
